refactor(phone-number-verification): inline modular entry into index
Migrate remaining e2e off global firebase; document app namespaced removal in v26 migration guide.
