+++ packages/phone-number-verification/typedoc.json
@@ -1,5 +1,5 @@
 {
   "$schema": "https://typedoc.org/schema.json",
-  "entryPoints": ["lib/modular.ts", "lib/types/pnv.ts"],
+  "entryPoints": ["lib/index.ts"],
   "tsconfig": "tsconfig.json"
 }
